backend/requirements.txt
# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Google Cloud / Vertex AI
google-genai>=1.0.0
google-cloud-aiplatform>=1.42.0
google-auth>=2.27.0

# XML Parsing
lxml>=5.1.0

# Graph/Network (DAG)
networkx>=3.2.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# ML (para templates e validação de imports)
scikit-learn>=1.3.0

# Retry Logic
tenacity>=8.2.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# WebSocket
websockets>=12.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0

